setup.py added; folders restructured
