+++ README.md
@@ -1,5 +1,5 @@
 # astrobject
-This module gather the basic tools one usually needs for astrophysics and cosmology usage. 
+This module gathers the basic tools one usually needs for astrophysics and cosmology usage. 
 
 Python
 ------
