+++ astrobject/astrobject/__init__.py
@@ -1,0 +1,22 @@
+#
+# Init the astro objects
+#
+
+# - fundamentals
+#import baseobject
+#import photometry 
+#import astrometry
+#import collection
+#import spectroscopy 
+# - directory of tools
+#import instruments
+#import collections
+
+#import transient # No used yet
+#import survey # No used yet
+
+
+from collections import *
+from instruments.instrument import *
+from photometry import *
+from baseobject import *
