+++ astrobject/astrobject/baseobject.py
