+++ astrobject/astrobject/collections/photospatial.py
