+++ astrobject/astrobject/instruments/baseinstrument.py
