+++ astrobject/astrobject/instruments/catalogues.py
