+++ astrobject/astrobject/instruments/instrument.py
