+++ astrobject/astrobject/instruments/ptf.py
