+++ astrobject/astrobject/instruments/stella.py
