+++ astrobject/astrobject/spectroscopy.py
