+++ astrobject/utils/decorators.py
