+++ astrobject/utils/shape.py
